% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdim.R
\name{read.dim}
\alias{read.dim}
\title{Import ESA SNAP .dim file}
\usage{
read.dim(dim_path)
}
\arguments{
\item{dim_path}{Path to the .dim file or a zip file contained .dim and .data.}
}
\description{
Import ESA SNAP .dim file
}
\examples{
library(raster)
library(ReadDIM)
im<-system.file("exd","td.dim",package = "ReadDIM")
read.dim(im)
}
\author{
Subhadip Datta
}
