% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdim.R
\name{read.dim}
\alias{read.dim}
\title{Import ESA SNAP .dim file}
\usage{
read.dim(dim_path, noDataV = NA, stacked = FALSE)
}
\arguments{
\item{dim_path}{Path to the .dim file.}

\item{noDataV}{Value of no data used in raster.}

\item{stacked}{If true create a stacked image else list}
}
\description{
Import ESA SNAP .dim file
}
\examples{
library(raster)
library(ReadDIM)
im<-system.file("exd","td.dim",package = "ReadDIM")
read.dim(im)
}
\author{
Subhadip Datta
}
